[ vm::al31fu ] rename `AL31F` to `AL31FM2`, and add `AL31FU`
