+++ pr47-core/raw_benches/bench_al31fm2_sync.rs
@@ -1,11 +1,11 @@
 use std::env;
 
 use pr47::data::Value;
-use pr47::vm::al31f::alloc::default_alloc::DefaultAlloc;
-use pr47::vm::al31f::compiled::CompiledProgram;
-use pr47::vm::al31f::exception::Exception;
-use pr47::vm::al31f::executor::vm_run_function_sync;
-use pr47::vm::al31f::test_program::{
+use pr47::vm::al31fm2::alloc::default_alloc::DefaultAlloc;
+use pr47::vm::al31fm2::compiled::CompiledProgram;
+use pr47::vm::al31fm2::exception::Exception;
+use pr47::vm::al31fm2::executor::vm_run_function_sync;
+use pr47::vm::al31fm2::test_program::{
     alloc_1m_program,
     bench_ffi_call_program,
     bench_ffi_call_program2,
