+++ pr47-core/src/lib.rs
@@ -27,8 +27,8 @@ pub mod vm;
 #[cfg(feature = "compiler")] pub mod sema;
 #[cfg(feature = "std47")]    pub mod std47;
 
-#[cfg(all(feature = "al31f-builtin-ops", not(feature = "al31f")))]
-compile_error!("using `al31f-builtin-ops` without `al31f` is meaningless");
+#[cfg(all(feature = "al31fm2-builtin-ops", not(feature = "al31fm2")))]
+compile_error!("using `al31fm2-builtin-ops` without `al31fm2` is meaningless");
 
 #[cfg(all(feature = "async-avoid-block", not(feature = "async")))]
 compile_error!("using `async-avoid-block` without `async` is meaningless");
@@ -43,4 +43,5 @@ compile_error!("features `async-astd` and `async-tokio` are mutually exclusive")
 compile_error!("feature `with-log` and `with-tracing` are mutually exclusive");
 
 #[cfg(test)]
+#[allow(unused_imports)]
 #[macro_use] extern crate variant_count;
