+++ pr47-core/src/vm/al31fm2/executor/checked_unary_ops.rs
