+++ pr47-core/src/vm/al31fm2/executor/coroutine_spawn.rs
@@ -6,13 +6,13 @@ use xjbutil::unchecked::{UncheckedSendFut, UncheckedSendSync};
 use crate::data::exception::{ExceptionInner, UncheckedException};
 use crate::data::Value;
 use crate::ffi::async_fn::{AsyncReturnType, Promise};
-use crate::vm::al31f::AL31F;
-use crate::vm::al31f::alloc::Alloc;
-use crate::vm::al31f::compiled::CompiledProgram;
-use crate::vm::al31f::exception::Exception;
-use crate::vm::al31f::executor::VMThread;
-use crate::vm::al31f::executor::{create_vm_child_thread, vm_thread_run_function};
-use crate::vm::al31f::stack::StackSlice;
+use crate::vm::al31fm2::AL31F;
+use crate::vm::al31fm2::alloc::Alloc;
+use crate::vm::al31fm2::compiled::CompiledProgram;
+use crate::vm::al31fm2::exception::Exception;
+use crate::vm::al31fm2::executor::VMThread;
+use crate::vm::al31fm2::executor::{create_vm_child_thread, vm_thread_run_function};
+use crate::vm::al31fm2::stack::StackSlice;
 
 #[cfg(feature = "async-astd")] use std::convert::Infallible as JoinError;
 #[cfg(feature = "async-astd")] use async_std::task::JoinHandle;
