+++ pr47-core/src/vm/al31fm2/executor/get_vm_makro.rs
