+++ pr47-core/src/vm/al31fm2/executor/impl_makro.rs
