+++ pr47-core/src/vm/al31fm2/executor/rtti.rs
