+++ pr47-core/src/vm/al31fm2/stack.rs
