+++ pr47-core/src/vm/mod.rs
@@ -1,2 +1,3 @@
-#[cfg(feature = "al31f")] pub mod al31f;
-#[cfg(feature = "d30f6")] pub mod d30f6;
+#[cfg(feature = "al31fm2")] pub mod al31fm2;
+#[cfg(feature = "al31fu")] pub mod al31fu;
+#[cfg(feature = "d30f6")]  pub mod d30f6;
